Add ML cut variation systematics
The ML cut variation systematics is implemented by deriving from the
Analyzer and can hence be used in the very same way:

mys_syst.step("syst_step")

and can also be wrapped into an AnalyzerManager

syst_mgr.analyze("syst_step")

which will run the step for all objects registered.

This class also makes use of the AnalyzerAfterBurner to run merging
steps after certain methods have been executed for all periods.
